+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - `com.facebook.bolts.AppLinkResolver` will only receive nonnull parameters and return nonnull values. `com.facebook.bolts.Capture` is removed.
+- Removed deprecated functions in `com.facebook.appevents.AppEventsLogger`
+- Removed deprecated function `ViewIndexer.sendToServerUnity(String)`
 
 ## [11.3.0] - 2021-09-06